// api/model/apiResponseDTOPouchListDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * POCHY API 명세서
 * POCHY Swagger UI
 * OpenAPI spec version: 1.0.0
 */
import type { PouchListDto } from './pouchListDto';

export interface ApiResponseDTOPouchListDto {
  code?: string;
  message?: string;
  result?: PouchListDto;
  success?: boolean;
}