// api/model/apiResponseDTOProfileDto.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * POCHY API 명세서
 * POCHY Swagger UI
 * OpenAPI spec version: 1.0.0
 */
import type { ProfileDto } from './profileDto';

export interface ApiResponseDTOProfileDto {
  code?: string;
  message?: string;
  result?: ProfileDto;
  success?: boolean;
}