// api/model/readListDto.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * POCHY API 명세서
 * POCHY Swagger UI
 * OpenAPI spec version: 1.0.0
 */

export interface ReadListDto {
  wishCosmeticsId?: number;
  brand?: string;
  productName?: string;
  category?: string;
  subCategory?: string;
  productImageUrl?: string;
}